nem-migrate-maybe-prepare-move: Move prepare_for_forward onto NgramEmbeddingManager (cut+paste)
Refactor chain ID: nem-migrate-maybe-prepare-move
